+++ src/models/campaign.js
